Time: 8 ms (45.9%), Space: 31.5 MB (76.45%) - LeetHub

+++ 0628-maximum-product-of-three-numbers/0628-maximum-product-of-three-numbers.cpp
@@ -3,16 +3,9 @@ public:
     int maximumProduct(vector<int>& nums) {
 
         sort(nums.begin(), nums.end());
-        long n = nums.size();
-        long option1 = nums[0] * nums[1] * nums[n-1];
-        long option2 = nums[n-1] * nums[n-2] * nums[n-3];
+        int n = nums.size();
+        int option1 = nums[0] * nums[1] * nums[n-1];
+        int option2 = nums[n-1] * nums[n-2] * nums[n-3];
         return max(option1, option2);
-
-        // sort(nums.begin(), nums.end());  //sort all
-        // long n = nums.size();
-        // long option1 = (long)nums[0]*nums[1]*nums[n-1];
-        // long option2 = (long)nums[n-1]*nums[n-2]*nums[n-3];
-        // long ans = max(option1, option2);
-        // return ans; 
     }
 };
